pyproject.toml
[project]
name = "GroceryTracker"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Rishabh Jain", email = "rishabjn10@gmail.com"},
]
dependencies = ["fastapi>=0.115.7", "isort>=5.13.2", "autopep8>=2.3.2", "flake8>=7.1.1", "black>=24.10.0", "SQLAlchemy>=2.0.37", "alembic>=1.14.1", "python-dotenv>=1.0.1", "pydantic-settings>=2.7.1", "psycopg2>=2.9.10", "psycopg2-binary>=2.9.10", "uvicorn>=0.34.0", "passlib[bcrypt]>=1.7.4", "python-jose[cryptography]>=3.3.0", "pydantic[email]>=2.10.6", "openai>=1.60.1", "python-multipart>=0.0.20"]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
